Implement 'generate' and 'validate' requests
The three endpoints are added:
 1. /licenses/generate
 2. /licenses/validate/license
 3. /licenses/validate/email

The '/licenses/generate' accepts email and returns the unique UUID of
the license number.
The '/licenses/validate/license' accepts email and license number then
returns true if such combination exists in database.
The '/licenses/validate/email' accepts email and returns true if it
exists in database.